Refactor minio client as agent
